fix  分钟格式化遇到22年超过int max的问题

+++ dashboard/src/main/java/com/jd/platform/hotkey/dashboard/common/domain/vo/HitCountVo.java
@@ -25,9 +25,9 @@ public class HitCountVo implements Serializable {
 
     private Integer hours;
 
-    private Integer minutes;
+    private Long minutes;
 
-    private Integer seconds;
+    private Long seconds;
 
     private Date createTime;
 
@@ -87,19 +87,19 @@ public class HitCountVo implements Serializable {
         this.hours = hours;
     }
 
-    public Integer getMinutes() {
+    public Long getMinutes() {
         return minutes;
     }
 
-    public void setMinutes(Integer minutes) {
+    public void setMinutes(Long minutes) {
         this.minutes = minutes;
     }
 
-    public Integer getSeconds() {
+    public Long getSeconds() {
         return seconds;
     }
 
-    public void setSeconds(Integer seconds) {
+    public void setSeconds(Long seconds) {
         this.seconds = seconds;
     }
 
@@ -125,8 +125,8 @@ public class HitCountVo implements Serializable {
         private BigDecimal ratio;
         private Integer days;
         private Integer hours;
-        private Integer minutes;
-        private Integer seconds;
+        private Long minutes;
+        private Long seconds;
         private Date createTime;
 
         private HitCountVoBuilder() {
@@ -167,12 +167,12 @@ public class HitCountVo implements Serializable {
             return this;
         }
 
-        public HitCountVoBuilder minutes(Integer minutes) {
+        public HitCountVoBuilder minutes(Long minutes) {
             this.minutes = minutes;
             return this;
         }
 
-        public HitCountVoBuilder seconds(Integer seconds) {
+        public HitCountVoBuilder seconds(Long seconds) {
             this.seconds = seconds;
             return this;
         }
